+++ lib/flutter_tyre.dart
@@ -3,11 +3,14 @@ library flutter_tyre;   //  ตั้งค่าให้เป็น Library  
 //  Import   Library  ที่ต้องการใช้
 
 import 'package:flutter/material.dart';
+
 // import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 
 //   ใช้  part    เพื่ออ้างอิงที่อยู่ของไฟล์ ที่ต้องการทั้งหมด
 
-part 'subwidgets/inkwell_text.dart';
+part 'widgets/text/text_title.dart';
+part 'widgets/text/text_body.dart';
+part 'widgets/appbar/app_bar_widget.dart';