+++ CHANGELOG.md
@@ -1,8 +1,16 @@
-## 0.0.1
 
-* TODO: Describe initial release.
 
-`This have Header Menu for convenient to make app web`
+
+
+## 0.0.6  
+
+* Remove: InkWellText() widget  
+* Add TextBody() widget  
+* Add TextTitle() widget  
+
+## 0.0.5
+
+* Add License  
 
 ## 0.0.3
 
@@ -10,5 +18,10 @@
 
 `InkwellText(text: "Somethings", onHover: (value) {}, onTap: () {}),`
 
-## 0.0.5
-* Add License
+## 0.0.1
+
+* TODO: Describe initial release.
+
+`Widget for flutter`  
+
+